LPS-140847 Avoid NPE on delete objectDefinition

+++ modules/apps/site-initializer/site-initializer-extender/site-initializer-extender-test/src/testIntegration/java/com/liferay/site/initializer/extender/internal/test/BundleSiteInitializerTest.java
@@ -157,8 +157,10 @@ public class BundleSiteInitializerTest {
 					serviceContext.getCompanyId(),
 					"C_TestBundleSiteInitializer");
 
-			_objectDefinitionLocalService.deleteObjectDefinition(
-				objectDefinition.getObjectDefinitionId());
+			if (objectDefinition != null) {
+				_objectDefinitionLocalService.deleteObjectDefinition(
+					objectDefinition.getObjectDefinitionId());
+			}
 		}
 
 		bundle.uninstall();
